+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix name and namespace regex used in create project from example command
+- Fix `compote autocomplete` incompatible with click 7.x issue
 
 ## [1.2.1] 2022-12-12